refactor(cli/core): delete legacy chat shim + replace_history + _chat_acp residue (R1 phase 6)
Final R1 cleanup. The cli/chat/sessions.py shim is gone — every caller
now imports from client.chat_sessions.X directly. Pure UI helpers moved
to cli/chat/_session_picker.py (ChatSessionListItem,
build_chat_session_list_items, resolve_chat_session_selector,
chat_session_to_legacy_dict). ChatSessions.replace_history deleted (zero
callers; upsert_with_history covers the real upsert case).

cli/chat/_chat_acp.py is gone. The single-approval modal helpers, the
session-allow cache (_session_approvals + get_session_approvals), the
_DecisionTuple shape, _map_decision_from_approval, and the pure
formatters (_format_permission_tool, _tool_action_key,
_stdio_is_interactive) all live in cli/chat/_permission_ui.py alongside
the existing PermissionUI class. Tests that previously imported
"_chat_acp as chat_acp_module" now import "_permission_ui as
chat_acp_module" — same surface, accurate name.

Server _chat_routes.py and the TUI orchestrator store keep the legacy
dict shape locally via chat_session_to_legacy_dict at the call boundary
rather than reaching back into a centralized shim. The server PATCH
handler was extracted into a module-level _patch_session helper to keep
_register_crud_routes under the McCabe-20 cap.

R1 is now complete. Manual UX smoke (CLI / TUI / web / VS Code) pending
before merge.

Net: ~-426 LOC across the cli + core packages. Test count: 1547 passed
(flat vs phase 5c, +12 new tests have not landed in this branch yet).
