+++ src/kagan/tui/screens/session_resume_modal.py
@@ -9,7 +9,10 @@ from textual.screen import ModalScreen
 from textual.widgets import Button, Footer, OptionList, Static
 from textual.widgets.option_list import Option
 
-from kagan.cli.chat.sessions import build_chat_session_list_items, list_chat_sessions
+from kagan.cli.chat._session_picker import (
+    build_chat_session_list_items,
+    chat_session_to_legacy_dict,
+)
 from kagan.core.errors import KaganError
 
 if TYPE_CHECKING:
@@ -64,7 +67,8 @@ class SessionResumeModal(ModalScreen[RecentSessionSelection | None]):
         try:
             projects = await self.kagan_app.core.projects.list()
             project_names = {project.id: project.name for project in projects}
-            sessions = await list_chat_sessions(self.kagan_app.core)
+            pairs = await self.kagan_app.core.chat_sessions.list_with_history()
+            sessions = [chat_session_to_legacy_dict(row, msgs) for row, msgs in pairs]
         except (KaganError, OSError, RuntimeError, ValueError):
             self.query_one("#session-resume-status", Static).update(
                 "Recent sessions are unavailable right now."
