+++ tests/server/test_chat_sse_persists_both_messages.py
@@ -18,7 +18,7 @@ from typing import TYPE_CHECKING, Any
 
 import pytest
 
-from kagan.cli.chat.sessions import get_chat_session
+from kagan.cli.chat._session_picker import chat_session_to_legacy_dict as _to_dict
 from kagan.core import KaganCore
 from kagan.server import _chat_routes
 from tests.helpers.chat_engine import ScriptedFactory
@@ -63,7 +63,8 @@ async def test_sse_turn_persists_both_user_and_assistant_rows(
         )
 
         session = await core.chat_sessions.create(source="web", label="t")
-        session_dict = await get_chat_session(core, session.id)
+        _pair = await core.chat_sessions.get_with_history(session.id)
+        session_dict = _to_dict(*_pair) if _pair is not None else None
         assert session_dict is not None
 
         ctx = SimpleNamespace(client=core)
